prototyp_global_coordinate/requirements.txt
#python==3.13.5
genesis-world==0.3.13
torch>=2.10.0
numpy>=2.3.0
PyYAML>=6.0
# rsl-rl-lib==2.2.4
rsl-rl-lib==5.0.1
tensorboard==2.20.0
